backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
google-genai
python-dotenv
gunicorn
groq
pdfplumber
